GH-1800: Add KafkaNullAwarePayloadArgumentResolver
Resolves https://github.com/spring-projects/spring-kafka/issues/1800

Move the code to a top-level class and document its use.
