+++ spring-kafka/src/main/java/org/springframework/kafka/annotation/KafkaListenerAnnotationBeanPostProcessor.java
@@ -64,7 +64,6 @@ import org.springframework.context.ApplicationContextAware;
 import org.springframework.context.ConfigurableApplicationContext;
 import org.springframework.context.expression.StandardBeanExpressionResolver;
 import org.springframework.core.MethodIntrospector;
-import org.springframework.core.MethodParameter;
 import org.springframework.core.OrderComparator;
 import org.springframework.core.Ordered;
 import org.springframework.core.annotation.AnnotatedElementUtils;
@@ -87,16 +86,12 @@ import org.springframework.kafka.retrytopic.RetryTopicBootstrapper;
 import org.springframework.kafka.retrytopic.RetryTopicConfiguration;
 import org.springframework.kafka.retrytopic.RetryTopicConfigurer;
 import org.springframework.kafka.retrytopic.RetryTopicInternalBeanNames;
-import org.springframework.kafka.support.KafkaNull;
 import org.springframework.kafka.support.TopicPartitionOffset;
 import org.springframework.lang.Nullable;
-import org.springframework.messaging.Message;
 import org.springframework.messaging.converter.GenericMessageConverter;
-import org.springframework.messaging.converter.MessageConverter;
 import org.springframework.messaging.converter.SmartMessageConverter;
 import org.springframework.messaging.handler.annotation.support.DefaultMessageHandlerMethodFactory;
 import org.springframework.messaging.handler.annotation.support.MessageHandlerMethodFactory;
-import org.springframework.messaging.handler.annotation.support.PayloadMethodArgumentResolver;
 import org.springframework.messaging.handler.invocation.HandlerMethodArgumentResolver;
 import org.springframework.messaging.handler.invocation.InvocableHandlerMethod;
 import org.springframework.util.Assert;
@@ -1108,36 +1103,6 @@ public class KafkaListenerAnnotationBeanPostProcessor<K, V>
 		@Override
 		public String getConversationId() {
 			return null;
-		}
-
-	}
-
-	private static class KafkaNullAwarePayloadArgumentResolver extends PayloadMethodArgumentResolver {
-
-		KafkaNullAwarePayloadArgumentResolver(MessageConverter messageConverter, Validator validator) {
-			super(messageConverter, validator);
-		}
-
-		@Override
-		public Object resolveArgument(MethodParameter parameter, Message<?> message) throws Exception { // NOSONAR
-			Object resolved = super.resolveArgument(parameter, message);
-			/*
-			 * Replace KafkaNull list elements with null.
-			 */
-			if (resolved instanceof List) {
-				List<?> list = ((List<?>) resolved);
-				for (int i = 0; i < list.size(); i++) {
-					if (list.get(i) instanceof KafkaNull) {
-						list.set(i, null);
-					}
-				}
-			}
-			return resolved;
-		}
-
-		@Override
-		protected boolean isEmptyPayload(Object payload) {
-			return payload == null || payload instanceof KafkaNull;
 		}
 
 	}
